requirements.txt
pandas>=2.2.0
openpyxl>=3.1.0
playwright>=1.45.0
PyYAML>=6.0.0
python-dotenv>=1.0.0
